Move withEnhancedLogger to api folder
